修改导出房屋信息接口，修改房屋实体字段类型

+++ common/src/main/java/com/jsy/community/qo/property/HouseQO.java
@@ -57,5 +57,5 @@ public class HouseQO implements Serializable {
     private Integer totalFloor;
     
     @ApiModelProperty(value = "楼宇分类")
-    private Integer buildingType;
+    private Long buildingType;
 }
